Updates to dependency management
