+++ java/gradle/src/main/kotlin/DependenciesTask.kt
@@ -1,9 +1,11 @@
 package org.processing.java.gradle
 
 import org.gradle.api.DefaultTask
-import org.gradle.api.GradleException
+import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.tasks.InputFile
+import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import java.io.File
 import java.io.ObjectInputStream
@@ -15,12 +17,16 @@ abstract class DependenciesTask: DefaultTask() {
     @InputFile
     val librariesMetaData: RegularFileProperty = project.objects.fileProperty()
 
+    @get:OutputDirectory
+    val outputJarsDirectory: DirectoryProperty = project.objects.directoryProperty()
+
     @InputFile
     val sketchMetaData: RegularFileProperty = project.objects.fileProperty()
 
     init{
         librariesMetaData.convention(project.layout.buildDirectory.file("processing/libraries"))
         sketchMetaData.convention(project.layout.buildDirectory.file("processing/sketch"))
+        outputJarsDirectory.convention(project.layout.buildDirectory.dir("processing/libs"))
     }
 
     @TaskAction
@@ -44,36 +50,21 @@ abstract class DependenciesTask: DefaultTask() {
                 library.jars.forEach { jar ->
                     jar.classes.forEach { className ->
                         if (className.startsWith(statement)) {
-                            dependencies.addAll(library.jars.map { it.path } )
+                            dependencies.addAll(library.jars.map { it.path })
                             return@import
                         }
                     }
                 }
             }
         }
-        project.dependencies.add("implementation",  project.files(dependencies) )
-
-        // TODO: Mutating the dependencies of configuration ':implementation' after it has been resolved or consumed. This
-
-        // TODO: Add only if user is compiling for P2D or P3D
-        // Add JOGL and Gluegen dependencies
-        project.dependencies.add("runtimeOnly", "org.jogamp.jogl:jogl-all-main:2.5.0")
-        project.dependencies.add("runtimeOnly", "org.jogamp.gluegen:gluegen-rt:2.5.0")
-
-        val os = System.getProperty("os.name").lowercase()
-        val arch = System.getProperty("os.arch").lowercase()
-
-        val variant = when {
-            os.contains("mac") -> "macosx-universal"
-            os.contains("win") && arch.contains("64") -> "windows-amd64"
-            os.contains("linux") && arch.contains("aarch64") -> "linux-aarch64"
-            os.contains("linux") && arch.contains("arm") -> "linux-arm"
-            os.contains("linux") && arch.contains("amd64") -> "linux-amd64"
-            else -> throw GradleException("Unsupported OS/architecture: $os / $arch")
+        // Copy discovered jars to the output directory so they can be wired into the
+        // compilation classpath at configuration time (Gradle 9 forbids mutating
+        // configurations after they have been observed/resolved).
+        val outputDir = outputJarsDirectory.get().asFile
+        outputDir.deleteRecursively()
+        outputDir.mkdirs()
+        dependencies.forEach { jar ->
+            jar.copyTo(outputDir.resolve(jar.name), overwrite = true)
         }
-
-        project.dependencies.add("runtimeOnly", "org.jogamp.gluegen:gluegen-rt:2.5.0:natives-$variant")
-        project.dependencies.add("runtimeOnly", "org.jogamp.jogl:nativewindow:2.5.0:natives-$variant")
-        project.dependencies.add("runtimeOnly", "org.jogamp.jogl:newt:2.5.0:natives-$variant")
     }
 }
